+++ kubejs/client_scripts/jei/chiselsandbits.js
@@ -1,0 +1,3 @@
+events.listen('jei.hide.items', function (event) {
+    event.hide('chiselsandbits:block_bit');
+})